Updated snippets/serializers.py, urls.py, views.py by adding a new route:
In order to add a new route that allows user to retrieve all of the
products / certificates of a specific farmer, the following updates
have been made :

- serializers.py : Added to FarmerSerializer an hyperlinkedfield and
a new field called 'certificate_products', that will redirect user to
all of the certificates / products data of the selected farmer.

- views.py : Created the CertificateProductList view, a DjangoRestMultipleModel's
library specific view, that will display the data of two combined
models/serializers (Product and Certificate). It's pagination is set up
by another new class : LimitPagination.

- urls.py : Added a path related to the the CertificateProductList view.